requirements.txt
annotated-types==0.6.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
emoji==2.9.0
fonttools==4.47.2
fpdf2==2.7.7
idna==3.6
inflect==7.0.0
iniconfig==2.0.0
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
pydantic_core==2.14.6
pydantic==2.5.3
pyfiglet==1.0.2
pytest==7.4.4
requests==2.31.0
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
validators==0.22.0
